processor/requirements.txt
paho-mqtt==2.1.0
psycopg[binary]==3.2.3
loguru==0.7.2
